Updated all the pages so non-admin features are hidden. In addition, fixed a bug with creating non-admin users.

+++ app/view/campaignList.tpl
@@ -7,16 +7,20 @@
 			echo "<span id='campaign_" . $c->id . "'>";
 			echo "<h1><a href=\"".BASE_URL."/campaigns/view/".$c->id."\">".$c->name." </a></h1>";
 			echo "<img class='campaign-thumbnail' src=\"https://upload.wikimedia.org/wikipedia/commons/thumb/6/6f/B-17F_formation_over_Schweinfurt%2C_Germany%2C_August_17%2C_1943.jpg/1200px-B-17F_formation_over_Schweinfurt%2C_Germany%2C_August_17%2C_1943.jpg\" alt=\"pic of campaign\" />";
+            if ($_SESSION['role'] == 0) {
 	    		echo "<button onclick='deleteCampaign(". $c->id .")'>X</button>";
+            }
 			echo "</span>";
 			echo "<br /><br />";
 		}
 	?>
 </div>
 <p>
+<?php if ($_SESSION['role'] == 0): ?>
 <div class="left-align">
 	<button id="add-campaign-button" class="large"> Add new Campaign</button>
 </div>
+<?php endif; ?>
 <div id="add-campaign-menu">
 
   <h1>Add Campaign</h1>
